requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
click-default-group==1.2.4
click-default-group-wheel==1.2.3
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.109.2
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==7.2.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-text-splitters==0.0.2
langgraph-sdk==0.3.9
llama-index==0.10.68
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llama_cloud==1.4.1
loguru==0.7.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.17.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.3
numpy==1.26.4
ollama==0.6.1
openai==1.109.1
orjson==3.11.7
ormsgpack==1.12.2
packaging==23.2
pandas==2.3.3
pillow==10.4.0
plotly==6.5.2
pluggy==1.6.0
propcache==0.4.1
protobuf==4.25.8
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.11.0
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
schedule==1.2.2
secure-smtplib==0.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.27
sqlite-fts4==1.0.3
sqlite-utils==3.35
starlette==0.36.3
streamlit==1.31.1
striprtf==0.0.26
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
twilio==9.10.2
typer==0.9.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.27.0.post1
validators==0.35.0
watchdog==6.0.0
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
